csv import
I’ve managed to do a very basic version of CSV import - for the time
being I’ve only managed to get the first name showing - need to figure
out how to concatenate it with surname. I also thought it would be nice
to have a company name underneath each name/last name as it could be
nicely styled in css, but that’s entirely optional. I’ve also fixed the
links in the sign up form so that it actually saves the details entered.